+++ src/main/java/com/liuzhihang/doc/view/utils/FeignPsiUtil.java
@@ -11,7 +11,6 @@ import com.intellij.psi.search.GlobalSearchScope;
 import com.liuzhihang.doc.view.constant.SpringConstant;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.Collection;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -35,11 +34,9 @@ public class FeignPsiUtil {
 
     public static List<PsiClass> findDocViewFromModule(Module module) {
 
-        Collection<PsiAnnotation> psiAnnotations = JavaAnnotationIndex.getInstance().get("FeignClient", module.getProject(), GlobalSearchScope.moduleScope(module));
-
         List<PsiClass> psiClasses = new LinkedList<>();
 
-        for (PsiAnnotation psiAnnotation : psiAnnotations) {
+        for (PsiAnnotation psiAnnotation : JavaAnnotationIndex.getInstance().getAnnotations("FeignClient", module.getProject(), GlobalSearchScope.moduleScope(module))) {
             PsiModifierList psiModifierList = (PsiModifierList) psiAnnotation.getParent();
             PsiElement psiElement = psiModifierList.getParent();