+++ src/main/java/com/liuzhihang/doc/view/utils/SpringPsiUtils.java
@@ -93,8 +93,8 @@ public class SpringPsiUtils extends ParamPsiUtils {
      */
     public static List<PsiClass> findDocViewFromModule(Module module) {
 
-        Collection<PsiAnnotation> psiAnnotations = JavaAnnotationIndex.getInstance().get("Controller", module.getProject(), GlobalSearchScope.moduleScope(module));
-        Collection<PsiAnnotation> restController = JavaAnnotationIndex.getInstance().get("RestController", module.getProject(), GlobalSearchScope.moduleScope(module));
+        Collection<PsiAnnotation> psiAnnotations = new ArrayList<>(JavaAnnotationIndex.getInstance().getAnnotations("Controller", module.getProject(), GlobalSearchScope.moduleScope(module)));
+        Collection<PsiAnnotation> restController = JavaAnnotationIndex.getInstance().getAnnotations("RestController", module.getProject(), GlobalSearchScope.moduleScope(module));
         psiAnnotations.addAll(restController);
         List<PsiClass> psiClasses = new LinkedList<>();